Reduce split communication size

+++ src/constant.rs
@@ -18,6 +18,6 @@ pub const COLS: usize = 7;
 // Number of rows for one hand
 pub const ROWS: usize = 5;
 
-pub const LEFT_DETECT_JUMPER_KEY: (u8, u8) = (2, 6);
+pub const LEFT_DETECT_JUMPER_KEY: (usize, usize) = (2, 6);
 
 pub const LAYER_NUM: usize = 2;
